version 0.11

+++ src/mechanics/Budget.java
@@ -2,7 +2,7 @@ package mechanics;
 
 public class Budget extends Statistics {
     @Override
-    public void change(int value) {
+    public void change(double value) {
         if(this.value + value <= 0) this.value = 0;
         this.value += value;
     }
