+++ src/mechanics/Ecology.java
@@ -4,7 +4,7 @@ import mechanics.consts.Values;
 
 public class Ecology extends Statistics {
     @Override
-    public void change(int value) {
+    public void change(double value) {
         if(this.value + value <= 0) this.value = 0;
         if(this.value + value <= Values.MAX_ECOLOGY.get()) this.value += value;
     }
