+++ src/mechanics/Statistics.java
@@ -3,13 +3,13 @@ package mechanics;
 import mechanics.consts.Values;
 
 public abstract class Statistics {
-    protected int value;
+    protected double value;
 
-    public int get() {
+    public double get() {
         return value;
     }
 
-    public abstract void change(int value);
+    public abstract void change(double value);
 
     protected Statistics() {
         this.value = Values.MAX_DEFAULT_VALUE.get();
