+++ src/mechanics/Tourism.java
@@ -5,7 +5,7 @@ import mechanics.consts.Values;
 public class Tourism extends Statistics {
 
     @Override
-    public void change(int value) {
+    public void change(double value) {
         if(this.value + value <= 0) this.value = 0;
         if(this.value + value <= Values.MAX_TOURISM.get()) this.value += value;
     }
